internal: move MIR generation DSL to `mirconstr` (#793)
## Summary

Move the routines and types that make up the DSL to the `mirconstr`
module and use the DSL in both the `injectdestructors` and `mirbridge`
module.

Apart from making the MIR code generation in both modules less complex,
manual, and error-prone, this is also significant progress towards the
MIR code coming out of the `injectdestructors` pass having proper type
information, which is a requirement for having MIR passes that run after
`injectdestructors`.

## Details

* move the `EValue` and `ChainEnd` types to `mirconstr`
* move the atomic DSL operands to `mirconstr`. The `notOp`, `modify`,
  `outOp`, and `tupleAccess` operands stay in `mirgen`
* add wrapper templates for DSL operands that `mirgen` still calls with
  a `TCtx` instance instead of a `MirNodeSeq`

As part of moving them, some operand routines that had a 'gen' prefix
are changed to not have it, making the routine names more consistent.

In addition, some cleanup in `mirconstr` is performed, and the DSL
extended a bit:
* `|=>`, `=>|`, and `previous` are removed. They were unused and didn't
  fit in with the overall design
* the `emit`, `symbol`, and `opParam` operands are added, all of which
  are only useful outside of `mirgen`

Another addition is the `predicate` meta operand, which makes it
possible to write chains where an operand is conditionally excluded:

```nim
chain(...): a() => predicate(cond) => b() => c()
```

Here, `b()` is only evaluated if `cond` evaluates to 'true'. Using two
chains plus an in-between 'if' statement was previously necessary to
achieve the same.

Finally, documentation about the DSL is added, and the
`injectdestructors` and `mirbridge` modules are changed to use the DSL
for generating MIR code. The logic from the `genInjectedSink` procedure
was previously duplicated into `genSinkFromTemporary`, but this is now
fixed.