internal: compiler trace free of legacy reports (#795)
## Summary

Legacy reports were being used to output compiler tracing, this was a
mistake, it made tracing slow and complicated, and added yet more
complexity to the legacy reports. This functionality moved from
`reports_sem` and other such modules into `debugutils`.

## Details

This is largely a lift-and-shift movement, data types and procedures are
moved over, chiefly `TraceSemReport` and `reportBody`.

Along with this move legacy reports cruft like `ReportLineInfo` and
inheriting from any legacy reports objects were removed. As well as a
key rename of `TraceSemReport` to `CompilerTrace`.

A large number of modules were impacted, thankfully mostly removing
dependencies. They already imported `debugutils` so the transition
simplifies consumers.

Finally, defaulted intermediate stacktraces between detailed traces as
the entire point of this is to see _what_ the compiler is doing, this
was a poor default.