internal: data-oriented `TLib` storage
Summary
=======

Replace storing `TLib` objects as `ref`s in `TSym` with storing all
`TLib` instances belonging to a module in a sequence that a `TSym` then
stores an index into.

This slightly reduces the size of `TSym` and, more importantly, makes
serializing the symbols and associated `TLib` instances easier: instead
of storing a `PackedLib` for *each* packed symbol, the packed symbols
now only store an index into the attached-to module's packed lib list.

Details
=======

For simplicity, the list with the libs for a module is stored in a
separate `seq` in `ModuleGraph`. The list cannot be stored in
`PContext`, as it needs to be accessible after semantic analysis, at
which point `PContext` is already gone.

Storing the `TLib` objects for a module happens when the module is
closed. While not strictly necessary at the moment, this is a
preparation for storing a `PSym` in `TLib.name` that is only set
up once the module is closed.

Aside from reducing the size of rodfiles, not storing a `PackedLib`
for each `PackedSym` also fixes another inefficiency, namely that
each deserialized got a unique, possibly duplicated, `TLib`
instance.