support foreign `TLib` references
A symbol can now reference the `TLib` belonging to a module that is not
the one the symbol is attached to. This works around an issue with
aliased structural types.

Also, use the ID provided by the module's ID generator instead of the
`skModule`s position. While both represent the same thing (a
``FileIndex``), querying the ID generator is more consistent with how it
works elsewhere.