internal: data-oriented storage for `TLib` (#797)
## Summary

Replace storing `TLib` objects as `ref`s in `TSym` with storing all
`TLib` instances belonging to a module in a sequence that a `TSym` then
stores an index-like reference into.

This makes serializing the symbols and associated `TLib` instances
easier: instead of storing a `PackedLib` for *each* packed symbol, the
packed symbols now only store a module+index tuple.

## Details

For simplicity, the list with the libs for a module is stored in a
separate `seq` in `ModuleGraph`. The list cannot be stored in
`PContext`, as it needs to be accessible after semantic analysis, at
which point `PContext` is already gone.

While the handle to the `TLib` instance stored with `TSym` would
ideally be a single index into the `TLib` list of the module a
symbol is attached to, the way symbols of aliased structural types
currently work (the aliased type's `TSym` content is partially
copied) makes this not possible, meaning that a symbol also has to
store the ID of the module the referenced `TLib` instance is part
of.

Storing the `TLib` objects for a module happens when the module is
closed. While not strictly necessary at the moment, this is a
preparation for storing a `PSym` in `TLib.name` that is only setup
when the module is closed.

Aside from reducing the size of rodfiles, not storing a `PackedLib`
for each `PackedSym` also fixes another inefficiency, namely that
each deserialized `TSym` got a unique, possibly duplicated, `TLib`
instance.