refactor(lexer): rename away from reports (#799)
## Summary

Rename `handleReport` to `handleLexerDiag` to differentiate it from the
rest of legacy reports.

## Details

Renamed `handleReport` to `handleLexerDiag`, this not only moves away
from 'reports' terminology, but makes it easier to differentiate this
proc from the rest of legacy reports. This is also more consistent with
the `parser.handleParserDiag`.

As part of this change two procs `newlineFollows` and `getIndentWidth`
are no longer exported as they have no public consumers.