nimscript: fix compiler crash with `.header` procs (#800)
## Summary

Fix the compiler crashing when running NimScript files that directly
contain procedure definitions using the `.header` or `.dynlib` pragma.

## Details

The module of a NimScript didn't use its own `IdGenerator`, but instead
re-used the `IdGenerator` instance belonging to the `ModuleGraph`. Apart
from leading to all symbols created as part of semantically analysing
the module being not attached to any module, this also caused an
`IndexDefect` when processing a `.header`/`.dynlib` pragma, as there the
module ID provided by the `IdGenerator` (which is -1 for the
`ModuleGraph`'s) is used as an index into to the `TLib` storage.

A dedicated `IdGenerator` is now created for a NimScript file, both
fixing the crash and guaranteeing that the module ID provided by a
`PContext`'s ID generator is a valid `FileIndex`.