contributing: update the commit guidelines
Summary
=======

Remove mentions of the outdated single commit requirement from both the
PR message template and `contributing.rst`. In addition, also add
details regarding what happens when merging the PR to both.

The intention with the latter is to make it easier for first-time
contributors to get an overview of the merge process, requirements, and
the role of the PR message.

Details
=======

With bors replaced with merge queues, the PR having a single commit is
no longer required, and manually squashing the commits is in fact often
times detrimental to understanding a merged PR. The section is replaced
with an encouragement to use multiple commits and a description of what
happens when the PR merged.

Rewriting history that has already been reviewed can make incremental
reviews harder (GitHub only shows the difference between the most
recent force push and the prior state), so an encouragement against
doing so is added to 'Anatomy of a Pull Request' section.

Since the line length requirement was only mentioned a single time in
the rather large `contributing.rst` document, it is not also mentioned
at the top of the PR message template.