vmgen: remove diagnostics representing internal errors
Summary
=======

Remove `vmgen`-related diagnostics that were about violations of
internal expectations/requirements. Detecting and reporting these kinds
of issues should not use diagnostics and reports, but rather `assert`,
`unreachable`, `internalAssert`, and `internalError`.

Details
=======

The removed internal diagnostics are `adVmGenNotAFieldSymbol`,
`adGenDiagInvalidObjectConstructor`, and
`adVmGenCodeGenGenericInNonMacro` (which was unused).

All `nkDotExpr` (not-a-field-symbol diagnostic) and `nkObjConstr`
(invalid-object-constructor diagnostic) trees reaching `vmgen` are
produced by `astgen`, meaning that if they have an unexpected structure,
`astgen` doesn't work as it should, which is a severe internal error.