vmjit: use MIR-based dependency discovery
Summary
=======

Move the JIT-related state from `TCtx` into a standalone type and
replace usage of the custom `PNode`-based dependency discovery with that
used by the backends.

The intention is to bring the processing in `vmjit` closer to that of
the backends (because `vmjit` can also be considered a backend) and to
prepare for the introduction of a dedicated code-generator/backend IR by
removing a usage of `PNode`. In addition, decoupling the JIT-state from
`TCtx` is part of the on-going project of splitting up `TCtx`.

For lifted globals used in code running at compile-time or in the
context of NimScript, there's a small difference in behaviour: they are
now reset to their default value when control-flow reaches the
definition statement, for now making them work like locals.

Details
=======

The key changes related to `TCtx`:
* introduce the `JitState` type, which represents the non-temporal JIT
  state. The compile-time evaluation context (`PEvalContext`) stores an
  instance of it
* remove `LinkState` and usages thereof; it is replaced by the
  `DiscoveryData` type also used by the backends
* as an interim solution until `vmbackend.nim` uses its own context
  type, add the `collectedGlobals` field to `TCtx` and replace usages of
  `LinkState` in `vmbackend` with it

The processing logic in `vmjit` now performs the following steps:
1. generate MIR code for the input
2. scan for definitions of globals and register them
3. rewrite definitions of globals into assignment
4. discover used constants and procedures

Apart from the missing handling of lifted globals, this is the same as
what the unified backend processing does. Step 2 and 4 were previously
implemented by the now-removed ``gatherDependencies``.

Before invoking `vmgen`, the newly discovered entities are registered
in `vmgen`'s link table, with the entities' index in the `DiscoveryData`
sequences being used as the link indices. In case of an error during
code generation, the `DiscoveryData` sequences are rolled back to their
state prior to code generation, which ensures that only entities used in
valid code are registered with `DiscoveryData`. Committing to the
entities only happens when they're loaded into the VM's execution
environment.

Rewriting the definitions of globals into assignments means that `var`
statements involving globals don't reach `vmgen` anymore and thus
that the corresponding logic in `vmgen` can be dropped.

Finally, the documentation of the `vmjit` module is updated to better
describe the module's purpose.