compiler: don't raise when NodeJS is missing
Summary
=======

Don't raise an exception when the NodeJS executable cannot be found.
In case that NodeJS is missing, this fixes:
- using the documentation generator with the JS backend crashing the
  compiler when a `runnableExample` is encountered
- `testament` aborting a category when attempting to run the first JS
  test

In addition, this also means that the compiler now exits with a proper
error instead of crashing because of an unhandled exception.

Details
=======

This is a revert of commit 0ae2d1ea88fdc7b6910ddbc960c7db086be383e4.
Except for one, all callsites of `findNodeJS` expect an empty string in
case a NodeJS installation cannot be found and use that information to
handle the error themselves: `docgen` handles it by not running
`runnableExamples` and `testament` treats and reports the affected test
as failed.

The only place where a missing NodeJS installation wasn't handled, and
what the reverted commit intended to fix, is the processing of the `run`
(or `-r`) option. There, the path is now verified to not be empty and an
error is reported (via `logError`) if it is.