support `uint64` for case statement operands (#820)
## Summary

Allow `uint64` values to be used as `case`-statement operands and fix
multiple compiler crashes with `uint` `of`-branch labels outside the
`int64` range. Previously, trying to use a `uint64` value with case
statements resulted in a "selector must be of an ordinal type" error.

* `uint64` values can now be used as `case`-statement operands (outside
  of objects)
* literal set constructors in `of`-branches don't crash the compiler
  when the range of the element value crosses `high(int64)`
* when using the C or JS backend, `..` ranges in `of`-branches don't
  crash the compiler when the lower bound is less-than-or-equal to,
  and the upper bound greater than `high(int64)`
* when using the JS backend, unsigned integer literals beyond
  `high(int64)` are not treated as `-1 - high(uint64) - value`
  anymore

## Details

`uint64` being disallowed as `case`-statement operands seems to have
been a leftover from when `uint64` was not considered an ordinal type.
Since it is considered an ordinal type now, not allowing it there is
inconsistent.

The compiler crashes were all caused by signed integer overflow
defects. All integer values are stored as `BiggestInt` (signed) in
`PNode`, even unsigned values. Prior to working with them, they have to
be either casted to an unsigned integer (in case the node represents
an unsigned value) or turned into an `Int128` value via `getOrdValue`/
`getInt` (both which take care of properly reinterpreting the value).

This was missing in `toTreeSet`, which is called when removing
duplicate elements from set literals used as `of`-branch labels.
Instead of turning the sets' lower bound (`firstOrd`) into a signed
integer, the value is now kept as an `Int128`, fixing the overflow
defect.

The other problems where with the C and JavaScript code generators,
both which were using the nodes `intVal` directly when iterating over
`nkRange`s -- they now use `Int128` values. In order to not having
to create a temporary `PNode`, the integer-literal rendering logic
for both code generators is moved to standalone procedures that take
an `Int128` value directly (which for the JavaScript backend fixes
large unsigned values being inverted).

Instead of inferring the integer type from the node kind, `cgen` now
requires all integer literals to be typed, which is preparation for the
code-generator IR.

When using the VM backend, `of`-branches with ranges crossing
`high(int64)` still don't work correctly, as the VM uses signed
integers for the comparisons. Fixing this requires a larger rethinking
of the VM's case statement support, and it is thus left as is for now.