add the code-generator IR
`astgen` is adjusted to produce `CgNode` instead of `PNode`. For this,
multiple `PNode` -> `CgNode` translation procedures had to be
introduced and `canUseView` + `flattenExpr` duplicated and adjusted for
`CgNode`. The general processing logic stays the same.

The module's document comment is also adjusted and an outdated mention
of "sections" (they are called "regions" in the MIR) fixed.

`astgen` as the name doesn't make much sense anymore and is going to be
changed to something more fitting.