compiler: introduce an IR for the code generators (#551)
## Summary

Add the `CgNode` intermediate representation, change `astgen` (which is
renamed to `cgirgen`) to output it instead of `PNode`, and update all
code generators to use `CgNode`. In order to keep the changes required
for the transition low, the differences of the new IR compared to
`PNode` are kept minimal.

The intent is to have an IR that can be evolved independently from sem
and the macro API. Only `PNode` is replaced so far, but both `PType`
and `PSym` are planned to also get a dedicated version for use by the
code generators. In addition, `PNode` can now be evolved without the
code generators having to be considered.

## Details

The core of the changes is the introduction of `CgNode`. It, for now,
also uses a `ref`-tree-based data-representation and is very similar,
in both structure and naming, to the `PNode`-subset previously used by
the code generators, but with some small simplifications/renamings
already applied where it makes sense.

**Naming differences:**
- `nkNimNodeLit` -> `cnkAstLit`
- `nkCurly` -> `cnkSetConstr`
- `nkBracket` -> `cnkArrayConstr`
- `nkClosure` -> `cnkClosureConstr`
- `nkHiddenDeref` -> `cnkDerefView`
- `nkHiddentStdConv` -> `cnkHiddenConv`
- `nkDiscardStmt` -> `cnkVoidStmt`
- `nkExprColonExpr` -> `cnkBinding`

**Structural differences:**
- the `while` statement is replaced with the `repeat` statement, since
  conditional loops (i.e. `while` statements) don't exist after the MIR
  phase. A `cnkRepeatStmt` node has a single sub-node, which is the
  loop's body
- there are no multi-branch `if` statements, since those don't exist
  during and after the MIR phase. A `cnkIfStmt` has the same structure
  as an `nkElifBranch`
- a `cnkReturnStmt` node has no sub nodes
- instead of `emit` being represented via a pragma statement, a
  dedicated statement (`cnkEmitStmt`) is used
- `cnkPragmaStmt`s don't have sub-nodes, but instead directly store the
  pragma's name
- `cnkObjConstr` doesn't have an extra type slot in the first position 
- there is only a single literal-node kind for the `int`, `uint`,
  `float`, and `string` types
- there are no dedicated nodes for `of` and `else` in `case` statements,
  both use `cnkBranch`
- there are no `nkVarSection|nkLetSection` + `nkIdentDefs` counterparts.
  Instead, definitions use the `cnkDef` node

Many more simplifications are possible, but they are left for future
PRs.

### Producing the IR

`astgen` is changed to produce `CgNode` instead of `PNode` IR. The
`canUseView` and `flattenExpr` procedures for `PNode` are copied and
adjusted for `CgNode`.

For literals, which are transported through the MIR phase as `PNode`s,
translation-to`CgNode` logic is added (`translateLit`). For integer-like
literals, whether the output node is an `cnkIntLit` or `cnkUIntLit` node
is decided based on the input node's *type*, not on its node *kind* --
this helps catching issues with incorrectly typed integer literals. For
`float32` literals, `translateLit` already narrows the value.

Since the module's name now doesn't apply anymore, it is changed to
`cgirnode`. In addition, the module is moved to the `backend` directory
and `generateAST` is renamed to `generateIR` -- the module's
documentation is also updated.

### Code generation

All three code generators and their surroundings are adjusted to operate
on `CgNode` instead of `PNode`, which for the most part means replacing
occurrences of `PNode` with `CgNode` and adjusting for the structural
difference listed above. Unrelated refactorings and clean-ups are kept
to a minimum.

There are two sources of `PNode`s still reaching into the code
generators: constants and type AST. When generating the code for
constants, their AST is first translated to `CgNode` IR before being
passed on to normal code generation, while for the type AST case,
special-purpose routines that still use `PNode` are added.

Routines still used by the code generators that are only available for
`PNode` are copied into the new `compat` module and adjusted to use
`CgNode`. This is meant to be an interim solution, and they are planned
to be phased out and removed again in the future.

Both the C and JavaScript code generation orchestrators passed the
`PNode` body of the main procedure directly to the code generator. This
doesn't work anymore, and it is replaced with canonicalizing (which
produces a `CgNode` tree) the body first.